Add registration, login, and logout functionality
Users can create an account by entering a username and password. Credentials are registered to a database. Connection and queries using PDO.

Users can login; code checks credentials against database to verify that the credentials are correct

Error handling messages added for invalid login credentials, taken username during login attempts, and locks the user out after 3 invalid attempts for 60 seconds.

When the user logs out, session variables are destroyed and the user is inauthenticated.

Website displays different header based on if user is logged in or not.